Revert "ratelimiter not working anymore, checking if i can get it working"
This reverts commit 0f94fb019999e24121cf9429cdab630da2e951df.
